refactor(api): move certification center serializer
